+++ intelmq/bots/collectors/http/collector_http.py
@@ -95,29 +95,34 @@ class HTTPCollectorBot(CollectorBot):
         self.logger.info("Report downloaded.")
 
         raw_reports = []
-        try:
-            zfp = zipfile.ZipFile(io.BytesIO(resp.content), "r")
-        except zipfile.BadZipfile:
-            raw_reports.append(resp.text)
+        if not self.extract_files:
+            try:
+                raw_reports = tuple(unzip(resp.content, True, try_gzip=False,
+                                          try_tar=False, logger=self.logger,
+                                          return_names=True))
+            except ValueError:
+                raw_reports.append((None, resp.text))
+            else:
+                self.logger.info('Extracting files: '
+                                 "'%s'.", "', '".join([file_name
+                                                       for file_name, _
+                                                       in raw_reports]))
         else:
-            self.logger.info('Extracting files:'
-                             "'%s'.", "', '".join(zfp.namelist()))
-            for filename in zfp.namelist():
-                raw_reports.append(zfp.read(filename))
-
-        if self.extract_files:
             if isinstance(self.extract_files, str) and len(self.extract_files):
                 self.extract_files = self.extract_files.split(",")
                 self.logger.info('Extracting files from archive: '
                                  "'%s'.", "', '".join(self.extract_files))
             else:
                 self.logger.info('Extracting all files from archive.')
-            raw_reports = [file for file in unzip(resp.content, self.extract_files)]
+            raw_reports = unzip(resp.content, self.extract_files,
+                                return_names=True, logger=self.logger)
 
-        for raw_report in raw_reports:
+        for file_name, raw_report in raw_reports:
             report = self.new_report()
             report.add("raw", raw_report)
             report.add("feed.url", http_url)
+            if file_name:
+                report.add("extra.file_name", file_name)
             self.send_message(report)